Style changes

+++ src/vw/Stereo/CostFunctions.h
@@ -39,19 +39,24 @@ namespace stereo {
   /// working type for accumulation operations or other intermediate
   /// results that require computational headroom.
   //
-  // The base type attempts to extract accumulator type from ImageView
+  // The base type attempts to extract accumulator type from ImageView. For bool and int
+  // types use an int64 accumulator to avoid overflow. All this logic is usually
+  // used with float images.
+  // TODO(oalexan1): The accumulators are best to be always float64.
   template <class ImageT>
   struct AbsAccumulatorType {
     typedef typename AbsAccumulatorType<typename ImageChannelType<ImageT>::type>::type type;
   };
-  template <class PixelT> struct AbsAccumulatorType<PixelMathBase<PixelT> > { typedef typename AbsAccumulatorType<typename CompoundChannelType<PixelT>::type>::type type; };
-  template <> struct AbsAccumulatorType<bool>        { typedef vw::int16   type; };
-  template <> struct AbsAccumulatorType<vw::uint8>   { typedef vw::int16   type; };
-  template <> struct AbsAccumulatorType<vw::int8>    { typedef vw::int16   type; };
-  template <> struct AbsAccumulatorType<vw::uint16>  { typedef vw::int32   type; };
-  template <> struct AbsAccumulatorType<vw::int16>   { typedef vw::int32   type; };
-  template <> struct AbsAccumulatorType<vw::uint32>  { typedef vw::int32   type; };
-  template <> struct AbsAccumulatorType<vw::int32>   { typedef vw::int32   type; };
+  template <class PixelT> struct AbsAccumulatorType<PixelMathBase<PixelT> > {
+    typedef typename AbsAccumulatorType<typename CompoundChannelType<PixelT>::type>::type type;
+  };
+  template <> struct AbsAccumulatorType<bool>        { typedef vw::int64   type; };
+  template <> struct AbsAccumulatorType<vw::uint8>   { typedef vw::int64   type; };
+  template <> struct AbsAccumulatorType<vw::int8>    { typedef vw::int64   type; };
+  template <> struct AbsAccumulatorType<vw::uint16>  { typedef vw::int64   type; };
+  template <> struct AbsAccumulatorType<vw::int16>   { typedef vw::int64   type; };
+  template <> struct AbsAccumulatorType<vw::uint32>  { typedef vw::int64   type; };
+  template <> struct AbsAccumulatorType<vw::int32>   { typedef vw::int64   type; };
   template <> struct AbsAccumulatorType<vw::uint64>  { typedef vw::int64   type; };
   template <> struct AbsAccumulatorType<vw::int64>   { typedef vw::int64   type; };
   template <> struct AbsAccumulatorType<vw::float32> { typedef vw::float32 type; };
@@ -62,10 +67,12 @@ namespace stereo {
   struct SqrDiffAccumulatorType {
     typedef typename SqrDiffAccumulatorType<typename CompoundChannelType<typename ImageT::pixel_type>::type>::type type;
   };
-  template <class PixelT> struct SqrDiffAccumulatorType<PixelMathBase<PixelT> > { typedef typename SqrDiffAccumulatorType<typename CompoundChannelType<PixelT>::type>::type type; };
-  template <> struct SqrDiffAccumulatorType<bool>        { typedef vw::int16   type; };
-  template <> struct SqrDiffAccumulatorType<vw::uint8>   { typedef vw::int32   type; };
-  template <> struct SqrDiffAccumulatorType<vw::int8>    { typedef vw::int32   type; };
+  template <class PixelT> struct SqrDiffAccumulatorType<PixelMathBase<PixelT> > {
+    typedef typename SqrDiffAccumulatorType<typename CompoundChannelType<PixelT>::type>::type type;
+  };
+  template <> struct SqrDiffAccumulatorType<bool>        { typedef vw::int64   type; };
+  template <> struct SqrDiffAccumulatorType<vw::uint8>   { typedef vw::int64   type; };
+  template <> struct SqrDiffAccumulatorType<vw::int8>    { typedef vw::int64   type; };
   template <> struct SqrDiffAccumulatorType<vw::uint16>  { typedef vw::int64   type; };
   template <> struct SqrDiffAccumulatorType<vw::int16>   { typedef vw::int64   type; };
   template <> struct SqrDiffAccumulatorType<vw::uint32>  { typedef vw::float32 type; };
@@ -79,21 +86,21 @@ namespace stereo {
   struct AbsDifferenceFunctor : BinaryReturn1stType {
     struct Helper : BinaryReturn1stType {
       template <class Arg1T, class Arg2T>
-      inline Arg1T operator()( Arg1T arg1, Arg2T arg2 ) const {
-        return ( arg1 < arg2 ) ? arg2 - arg1 : arg1 - arg2;
+      inline Arg1T operator()(Arg1T arg1, Arg2T arg2) const {
+        return (arg1 < arg2) ? arg2 - arg1 : arg1 - arg2;
       }
 
-      inline float operator()( float arg1, float arg2 ) const {
-        return fabs( arg1 - arg2 );
+      inline float operator()(float arg1, float arg2) const {
+        return fabs(arg1 - arg2);
       }
 
-      inline double operator()( double arg1, double arg2 ) const {
-        return fabs( arg1 - arg2 );
+      inline double operator()(double arg1, double arg2) const {
+        return fabs(arg1 - arg2);
       }
     };
 
     template <class Pixel1T, class Pixel2T>
-    Pixel1T operator()( Pixel1T const& arg1, Pixel2T const& arg2 ) const {
+    Pixel1T operator()(Pixel1T const& arg1, Pixel2T const& arg2) const {
       return BinaryCompoundFunctor<Helper,Pixel1T,Pixel2T>()(arg1, arg2);
     }
   };
@@ -102,38 +109,39 @@ namespace stereo {
     struct Helper : BinaryReturnUnaryTemplateBind1st<SqrDiffAccumulatorType> {
       template <class ArgT>
       inline typename SqrDiffAccumulatorType<ArgT>::type
-      operator()( ArgT const& arg1, ArgT const& arg2 ) const {
+      operator()(ArgT const& arg1, ArgT const& arg2) const {
         return (arg1 - arg2)*(arg1 - arg2);
       }
 
       inline SqrDiffAccumulatorType<uint8>::type
-      operator()( uint8 const& arg1, uint8 const& arg2 ) const {
-        return (arg1 < arg2 ) ? (arg2 - arg1)*(arg2 - arg1) :
+      operator()(uint8 const& arg1, uint8 const& arg2) const {
+        return (arg1 < arg2) ? (arg2 - arg1)*(arg2 - arg1) :
           (arg1 - arg2)*(arg1 - arg2);
       }
 
       inline SqrDiffAccumulatorType<uint16>::type
-      operator()( uint16 const& arg1, uint16 const& arg2 ) const {
-        return (arg1 < arg2 ) ? (arg2 - arg1)*(arg2 - arg1) :
+      operator()(uint16 const& arg1, uint16 const& arg2) const {
+        return (arg1 < arg2) ? (arg2 - arg1)*(arg2 - arg1) :
           (arg1 - arg2)*(arg1 - arg2);
       }
 
       inline SqrDiffAccumulatorType<uint32>::type
-      operator()( uint32 const& arg1, uint32 const& arg2 ) const {
-        return (arg1 < arg2 ) ? (arg2 - arg1)*(arg2 - arg1) :
+      operator()(uint32 const& arg1, uint32 const& arg2) const {
+        return (arg1 < arg2) ? (arg2 - arg1)*(arg2 - arg1) :
           (arg1 - arg2)*(arg1 - arg2);
       }
 
       inline SqrDiffAccumulatorType<uint64>::type
-      operator()( uint64 const& arg1, uint64 const& arg2 ) const {
-        return (arg1 < arg2 ) ? (arg2 - arg1)*(arg2 - arg1) :
+      operator()(uint64 const& arg1, uint64 const& arg2) const {
+        return (arg1 < arg2) ? (arg2 - arg1)*(arg2 - arg1) :
           (arg1 - arg2)*(arg1 - arg2);
       }
     };
 
     template <class Pixel1T, class Pixel2T>
-    typename PixelChannelCast<Pixel1T,typename SqrDiffAccumulatorType<typename PixelChannelType<Pixel1T>::type>::type>::type
-    operator()( Pixel1T const& arg1, Pixel2T const& arg2 ) const {
+    typename PixelChannelCast<Pixel1T, typename SqrDiffAccumulatorType
+                              <typename PixelChannelType<Pixel1T>::type>::type>::type
+    operator()(Pixel1T const& arg1, Pixel2T const& arg2) const {
       return BinaryCompoundFunctor<Helper,Pixel1T,Pixel2T>()(arg1, arg2);
     }
 
@@ -141,7 +149,8 @@ namespace stereo {
     template <class Args> struct result { typedef void type; };
     template <class FuncT, class Arg1T, class Arg2T>
     struct result<FuncT(Arg1T,Arg2T)> {
-      typedef typename PixelChannelCast<Arg1T,typename SqrDiffAccumulatorType<typename PixelChannelType<Arg1T>::type>::type>::type type;
+      typedef typename PixelChannelCast<Arg1T,
+        typename SqrDiffAccumulatorType<typename PixelChannelType<Arg1T>::type>::type>::type type;
     };
   };
 
@@ -149,7 +158,7 @@ namespace stereo {
     struct Helper : BinaryReturnUnaryTemplateBind1st<SqrDiffAccumulatorType> {
       template <class ArgT>
       inline typename boost::enable_if<boost::is_float<typename SqrDiffAccumulatorType<ArgT>::type>,typename SqrDiffAccumulatorType<ArgT>::type>::type
-      operator()( ArgT const& arg1, ArgT const& arg2 ) const {
+      operator()(ArgT const& arg1, ArgT const& arg2) const {
         return arg1 * arg2;
       }
 
@@ -158,14 +167,14 @@ namespace stereo {
       // is very hard to avoid overflowing.
       template <class ArgT>
       inline typename boost::disable_if<boost::is_float<typename SqrDiffAccumulatorType<ArgT>::type>,typename SqrDiffAccumulatorType<ArgT>::type>::type
-      operator()( ArgT const& arg1, ArgT const& arg2 ) const {
+      operator()(ArgT const& arg1, ArgT const& arg2) const {
         return typename SqrDiffAccumulatorType<ArgT>::type(arg1 * arg2) / 256;
       }
     };
 
     template <class Pixel1T, class Pixel2T>
     typename PixelChannelCast<Pixel1T,typename SqrDiffAccumulatorType<typename PixelChannelType<Pixel1T>::type>::type>::type
-    operator()( Pixel1T const& arg1, Pixel2T const& arg2 ) const {
+    operator()(Pixel1T const& arg1, Pixel2T const& arg2) const {
       return BinaryCompoundFunctor<Helper,Pixel1T,Pixel2T>()(arg1, arg2);
     }
 
@@ -206,11 +215,11 @@ namespace stereo {
     }
 
     // Does nothing
-    inline void cost_modification( ImageView<pixel_accumulator_type>& /*cost_metric*/,
-                                   Vector2i const& /*disparity*/ ) const {}
+    inline void cost_modification(ImageView<pixel_accumulator_type>& /*cost_metric*/,
+                                   Vector2i const& /*disparity*/) const {}
 
-    inline bool quality_comparison( accumulator_type cost,
-                                    accumulator_type quality ) const {
+    inline bool quality_comparison(accumulator_type cost,
+                                    accumulator_type quality) const {
       return cost < quality;
     }
   };
@@ -218,7 +227,8 @@ namespace stereo {
   template <class ImageT>
   struct SquaredCost {
     typedef typename SqrDiffAccumulatorType<ImageT>::type accumulator_type;
-    typedef typename PixelChannelCast<typename ImageT::pixel_type, accumulator_type>::type pixel_accumulator_type;
+    typedef typename PixelChannelCast<typename ImageT::pixel_type,
+                                      accumulator_type>::type pixel_accumulator_type;
 
     // Does nothing
     template <class ImageT1, class ImageT2>
@@ -235,20 +245,21 @@ namespace stereo {
     }
 
     // Does nothing
-    inline void cost_modification( ImageView<pixel_accumulator_type>& /*cost_metric*/,
-                                   Vector2i const& /*disparity*/ ) const {}
+    inline void cost_modification(ImageView<pixel_accumulator_type>& /*cost_metric*/,
+                                   Vector2i const& /*disparity*/) const {}
 
-    inline bool quality_comparison( accumulator_type cost,
-                                    accumulator_type quality ) const {
+    inline bool quality_comparison(accumulator_type cost,
+                                    accumulator_type quality) const {
       return cost < quality;
     }
   };
 
-  // Cross-correlation. Only float images are expected. 
+  // Cross-correlation. Only float images are expected.
   template <class ImageT>
   struct NCCCost {
     typedef typename SqrDiffAccumulatorType<ImageT>::type accumulator_type;
-    typedef typename PixelChannelCast<typename ImageT::pixel_type, accumulator_type>::type pixel_accumulator_type;
+    typedef typename PixelChannelCast<typename ImageT::pixel_type,
+                                      accumulator_type>::type pixel_accumulator_type;
     ImageView<pixel_accumulator_type> left_precision, right_precision;
 
     template <class ImageT1, class ImageT2>
@@ -269,14 +280,13 @@ namespace stereo {
       return result_type(left.impl(),right.impl());
     }
 
-    inline void cost_modification( ImageView<pixel_accumulator_type>& cost_metric,
-                                   Vector2i const& disparity ) const {
-      cost_metric *= sqrt( left_precision * crop(right_precision,
-                                                 bounding_box(left_precision)+disparity) );
+    inline void cost_modification(ImageView<pixel_accumulator_type>& cost_metric,
+                                   Vector2i const& disparity) const {
+      cost_metric *= sqrt(left_precision * crop(right_precision,
+                                                bounding_box(left_precision)+disparity));
     }
 
-    inline bool quality_comparison( accumulator_type cost,
-                                    accumulator_type quality ) const {
+    inline bool quality_comparison(accumulator_type cost, accumulator_type quality) const {
       return cost > quality;
     }
   };
